+++ src/parsers/contest/BeecrowdContestParser.ts
@@ -6,6 +6,6 @@ export class BeecrowdContestParser extends SimpleContestParser {
   protected problemParser = new BeecrowdProblemParser();
 
   public getMatchPatterns(): string[] {
-    return ['https://www.beecrowd.com.br/judge/*/challenges/contest/*'];
+    return ['https://judge.beecrowd.com/*/challenges/contest/*'];
   }
 }